refactor: consolidate backtesting modules
